04-deal-memo-generator-done/pyproject.toml
[project]
name = "deal-memo-generator-done"
version = "0.1.0"
description = "Part 4: Japanese Deal Memo Generator (Complete)"
requires-python = ">=3.12"
dependencies = [
    "python-dotenv>=1.0.0",
    "sap-ai-sdk-gen[langchain]>=1.0.0",
    "langchain-hana>=0.1.0",
    "langgraph>=0.2.0",
    "hdbcli>=2.20.0",
    "yfinance>=0.2.40",
]

[tool.uv]
dev-dependencies = []
